+++ lib/aliyun/log/record/exception.rb
@@ -1,0 +1,9 @@
+module Aliyun
+  module Log
+    module Record
+      class ArgumentError < StandardError; end
+      class UnknownAttributeError < StandardError; end
+      class ProjectNameError < StandardError; end
+    end
+  end
+end